Rename items files to have more accurate names
